INS-593: move api helpers to separate file
